Scale lane and move ball
The original plans is 10 m by 10 m. A bowling lane is roughly 20 m long
and 1 m wide. The plane is scaled to that dimension.

The ball is moved to the beginning of the lane and made the correct
size, about 20 cm diameter.

To make sure the ball is visible at the beginning of the lane, the
camera is moved back by 2 m.